feat(scripts): admin child registration safe payload
