+++ packages/scripts/src/migration/dvm2-designated-voting/common.ts
@@ -17,28 +17,6 @@ export interface DesignatedVotingOwnerWithBalance {
   balance: BigNumber;
 }
 
-export interface GnosisPayload {
-  version: string;
-  chainId: string;
-  createdAt: number;
-  meta: {
-    name: string;
-    description: string;
-    txBuilderVersion: string;
-    createdFromSafeAddress: string;
-    createdFromOwnerAddress: string;
-    checksum: string;
-  };
-  transactions: GnosisTransaction[];
-}
-
-export interface GnosisTransaction {
-  to: string;
-  value: string;
-  contractMethod: { inputs: { internalType: string; name: string; type: string }[]; name: string; payable: boolean };
-  contractInputsValues: { [key: string]: string };
-}
-
 export async function getDesignatedVotingContractsOwnedByOwner(
   owner: string
 ): Promise<DesignatedVotingOwnerWithBalance[]> {
@@ -100,42 +78,4 @@ async function _runEtherscanApiCall(url: string): Promise<EtherscanTransactionLi
     headers: { Accept: "application/json", "Content-Type": "application/json" },
   });
   return (await response.json()).result;
-}
-
-export function baseSafePayload(
-  chainId: number,
-  name: string,
-  description: string,
-  createdFromSafeAddress: string
-): GnosisPayload {
-  return {
-    version: "1.0",
-    chainId: chainId.toString(),
-    createdAt: Date.now(),
-    meta: {
-      name,
-      description,
-      txBuilderVersion: "1.13.2",
-      createdFromSafeAddress,
-      createdFromOwnerAddress: "",
-      checksum: "",
-    },
-    transactions: [],
-  };
-}
-
-export function appendTxToSafePayload(
-  payload: any,
-  to: string,
-  contractMethod: any,
-  contractInputsValues: any
-): GnosisPayload {
-  payload.transactions.push({
-    to,
-    value: "0",
-    data: null,
-    contractMethod,
-    contractInputsValues,
-  });
-  return payload;
 }
